+++ Backend/internal/core/service/room.go
@@ -1,11 +1,12 @@
 package service
 
 import (
-	"context"
 	"time"
 
 	"github.com/Coke3a/HotelManagement/internal/core/domain"
 	"github.com/Coke3a/HotelManagement/internal/core/port"
+	"github.com/gin-gonic/gin"
+	"log/slog"
 )
 
 type RoomService struct {
@@ -20,7 +21,7 @@ func NewRoomService(repo port.RoomRepository, logRepo port.LogRepository) *RoomS
 	}
 }
 
-func (rs *RoomService) RegisterRoom(ctx context.Context, room *domain.Room) (*domain.Room, error) {
+func (rs *RoomService) RegisterRoom(ctx *gin.Context, room *domain.Room) (*domain.Room, error) {
 	// Basic validation
 	if room.RoomNumber == "" || room.TypeID == 0 {
 		return nil, domain.ErrInvalidData
@@ -33,10 +34,27 @@ func (rs *RoomService) RegisterRoom(ctx context.Context, room *domain.Room) (*do
 		}
 		return nil, domain.ErrInternal
 	}
+
+	userID, exists := ctx.Get("userID")
+	if !exists {
+		return nil, domain.ErrUnauthorized
+	}
+	// Create a log
+	log := &domain.Log{
+		RecordID:  room.ID,
+		Action:    "CREATE",
+		UserID:    userID.(uint64),
+		TableName: "rooms",
+	}
+	_, err = rs.logRepo.CreateLog(ctx, log)
+	if err != nil {
+		slog.Error("Error creating log", "error", err)
+	}
+
 	return room, nil
 }
 
-func (rs *RoomService) GetRoom(ctx context.Context, id uint64) (*domain.Room, error) {
+func (rs *RoomService) GetRoom(ctx *gin.Context, id uint64) (*domain.Room, error) {
 	room, err := rs.repo.GetRoomByID(ctx, id)
 	if err != nil {
 		if err == domain.ErrDataNotFound {
@@ -48,7 +66,7 @@ func (rs *RoomService) GetRoom(ctx context.Context, id uint64) (*domain.Room, er
 	return room, nil
 }
 
-func (rs *RoomService) ListRooms(ctx context.Context, skip, limit uint64) ([]domain.Room, error) {
+func (rs *RoomService) ListRooms(ctx *gin.Context, skip, limit uint64) ([]domain.Room, error) {
 	rooms, err := rs.repo.ListRooms(ctx, skip, limit)
 	if err != nil {
 		return nil, domain.ErrInternal
@@ -57,7 +75,7 @@ func (rs *RoomService) ListRooms(ctx context.Context, skip, limit uint64) ([]dom
 	return rooms, nil
 }
 
-func (rs *RoomService) UpdateRoom(ctx context.Context, room *domain.Room) (*domain.Room, error) {
+func (rs *RoomService) UpdateRoom(ctx *gin.Context, room *domain.Room) (*domain.Room, error) {
 	existingRoom, err := rs.repo.GetRoomByID(ctx, room.ID)
 	if err != nil {
 		if err == domain.ErrDataNotFound {
@@ -91,10 +109,26 @@ func (rs *RoomService) UpdateRoom(ctx context.Context, room *domain.Room) (*doma
 		return nil, domain.ErrInternal
 	}
 
+	userID, exists := ctx.Get("userID")
+	if !exists {
+		return nil, domain.ErrUnauthorized
+	}
+	// Create a log
+	log := &domain.Log{
+		RecordID:  room.ID,
+		Action:    "UPDATE",
+		UserID:    userID.(uint64),
+		TableName: "rooms",
+	}
+	_, err = rs.logRepo.CreateLog(ctx, log)
+	if err != nil {
+		slog.Error("Error creating log", "error", err)
+	}
+
 	return room, nil
 }
 
-func (rs *RoomService) DeleteRoom(ctx context.Context, id uint64) error {
+func (rs *RoomService) DeleteRoom(ctx *gin.Context, id uint64) error {
 	_, err := rs.repo.GetRoomByID(ctx, id)
 	if err != nil {
 		if err == domain.ErrDataNotFound {
@@ -103,10 +137,26 @@ func (rs *RoomService) DeleteRoom(ctx context.Context, id uint64) error {
 		return domain.ErrInternal
 	}
 
+	userID, exists := ctx.Get("userID")
+	if !exists {
+		return domain.ErrUnauthorized
+	}
+	// Create a log
+	log := &domain.Log{
+		RecordID:  id,
+		Action:    "DELETE",
+		UserID:    userID.(uint64),
+		TableName: "rooms",
+	}
+	_, err = rs.logRepo.CreateLog(ctx, log)
+	if err != nil {
+		slog.Error("Error creating log", "error", err)
+	}
+
 	return rs.repo.DeleteRoom(ctx, id)
 }
 
-func (rs *RoomService) GetAvailableRooms(ctx context.Context, checkInDate, checkOutDate time.Time) ([]domain.RoomWithRoomType, error) {
+func (rs *RoomService) GetAvailableRooms(ctx *gin.Context, checkInDate, checkOutDate time.Time) ([]domain.RoomWithRoomType, error) {
 	if checkInDate.After(checkOutDate) {
 		return nil, domain.ErrInvalidData
 	}
@@ -119,7 +169,7 @@ func (rs *RoomService) GetAvailableRooms(ctx context.Context, checkInDate, check
 	return rooms, nil
 }
 
-func (rs *RoomService) ListRoomsWithRoomType(ctx context.Context, skip, limit uint64) ([]domain.RoomWithRoomType, error) {
+func (rs *RoomService) ListRoomsWithRoomType(ctx *gin.Context, skip, limit uint64) ([]domain.RoomWithRoomType, error) {
 	rooms, err := rs.repo.ListRoomsWithRoomType(ctx, skip, limit)
 	if err != nil {
 		return nil, domain.ErrInternal
